Counts can be too big for i8

+++ src/main.rs
@@ -219,7 +219,7 @@ fn fetch_comments(url: hyper::Uri) -> impl Future<Item=(Vec<Comment>, String), E
 struct Story {
     title: String,
     created_at: String,
-    score: i8,
+    score: i16,
     comment_count: u8,
     short_id: String,
     short_id_url: String,
@@ -243,7 +243,7 @@ struct CommentRoot {
 struct Comment {
     comment: String,
     created_at: String,
-    score: i8,
+    score: i16,
     indent_level: u8,
     commenting_user: User,
 }
